Make noarch and remove poetry from run requirements
As far as I can tell the package is purely typing stubs and thus should
be suitable for a noarch package, furthermore the wheel on pypi is also
noarch. Instead I introduced a python >=3.8 version constraint to
match the previous skip constraint.

Poetry was added in #95 as both a build, host and run dependency. As far
as I can tell it should not be necessary to add it as a runtime
dependency and it pulls in a lot of other dependencies when installed so
if possible it would be nice to avoid having it as a runtime dependency.